refactor : list 디렉토리로 페이지네이션 리스트 형식, 위치 변경
